+++ subprojects/testing-jvm/src/main/java/org/gradle/api/tasks/testing/Test.java
@@ -35,7 +35,7 @@ import org.gradle.api.internal.tasks.testing.DefaultTestTaskReports;
 import org.gradle.api.internal.tasks.testing.NoMatchingTestsReporter;
 import org.gradle.api.internal.tasks.testing.TestFramework;
 import org.gradle.api.internal.tasks.testing.TestResultProcessor;
-import org.gradle.api.internal.tasks.testing.TestWorkerProgressListener;
+import org.gradle.api.internal.tasks.testing.worker.TestWorkerProgressListener;
 import org.gradle.api.internal.tasks.testing.detection.DefaultTestExecuter;
 import org.gradle.api.internal.tasks.testing.detection.TestExecuter;
 import org.gradle.api.internal.tasks.testing.filter.DefaultTestFilter;
